requirements.txt
telethon>=1.28.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
googletrans==4.0.0rc1
